myblog/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.4.1
diff-match-patch==20181111
Django==2.0
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-js-asset==1.1.0
django-reversion==3.0.2
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.7
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.10
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.7
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0
